BE/feat: 결제/출금/입금 기능 구현

+++ Backend/src/main/java/com/ts/taesan/global/openfeign/bank/dto/request/ChargeRequest.java
@@ -16,6 +16,6 @@ public class ChargeRequest {
     private String senderAccNum;
 
     @JsonProperty("trans_amt")
-    private Integer transAmt;
+    private Long transAmt;
 
 }
